music-recommender/requirements.txt
pandas
numpy
scikit-learn
hdbscan
faiss-cpu
google-cloud-storage
google-cloud-bigquery
sqlalchemy
psycopg2-binary
python-dotenv
